backend/requirements.txt
alembic==1.9.2
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
databases==0.7.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
fastapi-jwt-auth==0.5.0
flatbuffers==23.1.21
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.2
mpmath==1.2.1
numpy==1.24.1
onnx==1.13.0
onnxruntime==1.13.1
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pydantic==1.10.4
PyJWT==1.7.1
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
sympy==1.11.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.11.0
